refactor(ensadmin): full integration of Indexing Status API
