+++ packages/ensnode-sdk/src/ensindexer/indexing-status/zod-schemas.ts
@@ -169,9 +169,10 @@ export const makeChainIndexingCompletedStatusSchema = (valueLabel: string = "Val
       },
     )
     .refine(
-      ({ config, latestIndexedBlock }) => blockRef.isEqualTo(latestIndexedBlock, config.endBlock),
+      ({ config, latestIndexedBlock }) =>
+        blockRef.isBeforeOrEqualTo(latestIndexedBlock, config.endBlock),
       {
-        error: `latestIndexedBlock must be the same as config.endBlock.`,
+        error: `latestIndexedBlock must be before or same as config.endBlock.`,
       },
     );
